Move more code into command_line/utils
